packages/feature-store/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepsynaps-feature-store"
version = "0.1.0"
description = "DeepSynaps Layer 2 Feature Store (Feast + Redis + streaming workers)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.11.0,<3.0.0",
  "pytest>=8.0.0,<9.0.0",
  "redis>=5.0.0,<6.0.0",
  "faust-streaming>=0.11.0,<0.12.0",
  "feast>=0.45.0,<0.48.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0,<1.0.0",
  "mypy>=1.11.0,<2.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

